Object Store Clean Shutdown (#15)
* Startup replication service without crash.

* Add back unit test runs.

* No test target currently.

* obj_store: graceful shutdown.

* Seperate common procedures to common object.
